Acomodamos los tabs
