+++ src/app/pages/explore-container/explore-container.component.scss
