+++ src/app/pages/explore-container/explore-container.module.ts
