+++ src/app/pages/tab1/tab1.module.ts
@@ -2,10 +2,10 @@ import { IonicModule } from '@ionic/angular';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { Tab3Page } from './tab3.page';
-import { ExploreContainerComponentModule } from '../explore-container/explore-container.module';
+import { Tab1Page } from './tab1.page';
 
-import { Tab3PageRoutingModule } from './tab3-routing.module';
+import { Tab1PageRoutingModule } from './tab1-routing.module';
+import { ExploreContainerComponentModule } from '../explore-container/explore-container.module';
 
 @NgModule({
   imports: [
@@ -13,8 +13,8 @@ import { Tab3PageRoutingModule } from './tab3-routing.module';
     CommonModule,
     FormsModule,
     ExploreContainerComponentModule,
-    Tab3PageRoutingModule
+    Tab1PageRoutingModule
   ],
-  declarations: [Tab3Page]
+  declarations: [Tab1Page]
 })
-export class Tab3PageModule {}
+export class Tab1PageModule {}
