+++ src/app/pages/tab2/tab2-routing.module.ts
