+++ src/app/pages/tab2/tab2.module.ts
@@ -2,10 +2,10 @@ import { IonicModule } from '@ionic/angular';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { Tab1Page } from './tab1.page';
-import { ExploreContainerComponentModule } from '../explore-container/explore-container.module';
+import { Tab2Page } from './tab2.page';
 
-import { Tab1PageRoutingModule } from './tab1-routing.module';
+import { Tab2PageRoutingModule } from './tab2-routing.module';
+import { ExploreContainerComponentModule } from '../explore-container/explore-container.module';
 
 @NgModule({
   imports: [
@@ -13,8 +13,8 @@ import { Tab1PageRoutingModule } from './tab1-routing.module';
     CommonModule,
     FormsModule,
     ExploreContainerComponentModule,
-    Tab1PageRoutingModule
+    Tab2PageRoutingModule
   ],
-  declarations: [Tab1Page]
+  declarations: [Tab2Page]
 })
-export class Tab1PageModule {}
+export class Tab2PageModule {}
