+++ src/app/pages/tabs/tabs.module.ts
