+++ src/app/pages/tabs/tabs.page.html
@@ -1,0 +1,15 @@
+<ion-tabs>
+
+  <ion-tab-bar slot="bottom">
+    <ion-tab-button tab="tab1">
+      <ion-icon name="list-outline"></ion-icon>
+      <ion-label>Earrings</ion-label>
+    </ion-tab-button>
+
+    <ion-tab-button tab="tab2">
+      <ion-icon name="checkmark-done-outline"></ion-icon>
+      <ion-label>Finished</ion-label>
+    </ion-tab-button>
+  </ion-tab-bar>
+
+</ion-tabs>
