+++ src/app/pages/tabs/tabs.page.spec.ts
